### 2024-12-11 #7583 @editorialbot
[PRE REVIEW]: unxt: A Python package for unit-aware computing with JAX
**Submitting author:** @nstarman (<a href="http://orcid.org/0000-0003-3954-3291">Nathaniel Starkman</a>)
**Repository:** https://github.com/GalacticDynamics/unxt
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
