### 2024-12-11 17a0ab6
Backport PR #313: chore: update pre-commit hooks (#323)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
