+++ src/unxt/_src/dimensions/__init__.py
@@ -4,6 +4,6 @@ This is the private implementation of the dimensions module.
 
 """
 
-__all__ = ["dimension", "dimension_of", "AbstractDimension"]
+__all__ = ["AbstractDimension", "dimension", "dimension_of"]
 
 from .core import AbstractDimension, dimension, dimension_of
