+++ .pre-commit-config.yaml
@@ -1,17 +1,18 @@
 ci:
   autoupdate_schedule: "monthly"
-  autoupdate_commit_msg: "chore: update pre-commit hooks"
-  autofix_commit_msg: "style: pre-commit fixes"
+  autoupdate_commit_msg: "👷: update pre-commit hooks"
+  autofix_commit_msg: "🎨: pre-commit fixes"
 
 default_stages: [pre-commit, pre-push]
 
 repos:
   - repo: https://github.com/commitizen-tools/commitizen
-    rev: v3.30.1
+    rev: v4.0.0
     hooks:
       - id: commitizen
-      # - id: commitizen-branch
-      #   stages: [push]
+        args:
+          - "--"
+          - "--config-file=.czrc"
 
   - repo: meta
     hooks:
@@ -44,14 +45,14 @@ repos:
       - id: rst-inline-touching-normal
 
   - repo: https://github.com/python-jsonschema/check-jsonschema
-    rev: 0.29.4
+    rev: 0.30.0
     hooks:
       - id: check-dependabot
       - id: check-github-workflows
       - id: check-readthedocs
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.7.3"
+    rev: "v0.8.1"
     hooks:
       # Run the linter
       - id: ruff
@@ -68,7 +69,7 @@ repos:
         additional_dependencies: [black==23.*]
 
   - repo: https://github.com/rbubley/mirrors-prettier
-    rev: "v3.3.3"
+    rev: "v3.4.1"
     hooks:
       - id: prettier
         types_or: [yaml, markdown, html, css, scss, javascript, json]
