### 2024-12-11 bd91bd2
Backport PR #317: build(deps): bump the actions group with 2 updates (#325)
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
