### 2024-12-11 c3e33a1
build(deps): bump actions/attest-build-provenance in the actions group (#322)
Bumps the actions group with 1 update: [actions/attest-build-provenance](https://github.com/actions/attest-build-provenance).


Updates `actions/attest-build-provenance` from 2.0.1 to 2.1.0
- [Release notes](https://github.com/actions/attest-build-provenance/releases)
- [Changelog](https://github.com/actions/attest-build-provenance/blob/main/RELEASE.md)
- [Commits](https://github.com/actions/attest-build-provenance/compare/v2.0.1...v2.1.0)

---
updated-dependencies:
- dependency-name: actions/attest-build-provenance

[…]
