+++ .github/workflows/cd.yml
@@ -48,7 +48,7 @@ jobs:
           path: dist
 
       - name: Generate artifact attestation for sdist and wheel
-        uses: actions/attest-build-provenance@v2.0.1
+        uses: actions/attest-build-provenance@v2.1.0
         with:
           subject-path: "dist/*"
 
@@ -77,7 +77,7 @@ jobs:
           path: dist
 
       - name: Generate artifact attestation for sdist and wheel
-        uses: actions/attest-build-provenance@v2.0.1
+        uses: actions/attest-build-provenance@v2.1.0
         with:
           subject-path: "dist/*"
 
