### 2024-12-11 860f0ee
🔧 config(uv): add temporary jax constraint (#321) (#326)
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
(cherry picked from commit 9fb427b40d93915f208219846d35c5db0f9db8a3)
