### 2024-12-11 eda005c
Backport PR #319: Update CI badge in README.md (#329)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
