### 2024-12-11 0dcadea
Backport PR #318: 🩹 fix-simple(quantity): remove astropy v6 repr code (#330)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
