### 2024-12-11 d3015da
Backport PR #320: ⬆️ dep-bump(quaxed): v0.7.0+
