### 2024-12-16 be5b085
🔧 config(ruff): ignore TC003
