+++ pyproject.toml
@@ -189,6 +189,7 @@
       "PLR2004", # Magic value used in comparison
       "PYI041",  # Use `complex` instead of `int | complex` <- plum is more strict
       "RUF022",  # `__all__` is not sorted
+      "TC003",   # typing-only-standard-library-import
       "TD002",   # Missing author in TODO
       "TD003",   # Missing issue link on the line following this TODO
     ]
