### 2024-12-17 b867ef8
🔧 config(codecov): exclude TYPE_CHECKING blocks
