+++ pyproject.toml
@@ -128,7 +128,11 @@
 
 
 [tool.coverage]
-  report.exclude_also = ['\.\.\.', 'if typing.TYPE_CHECKING:']
+  report.exclude_also = [
+    '\.\.\.',
+    'if typing.TYPE_CHECKING:',
+    'if TYPE_CHECKING:',
+  ]
   run.source          = ["unxt"]
   run.branch          = true
 
