### 2024-12-17 6606824
🔧 config(ruff): ignore TC003 and associated fixes (#332)
* 🔧 config(ruff): ignore TC003
* 🏷️ types(quantity): improve PEP 563 annotations in quantity mixin
* 🔧 config(codecov): exclude TYPE_CHECKING blocks
