+++ pyproject.toml
@@ -128,7 +128,11 @@
 
 
 [tool.coverage]
-  report.exclude_also = ['\.\.\.', 'if typing.TYPE_CHECKING:']
+  report.exclude_also = [
+    '\.\.\.',
+    'if typing.TYPE_CHECKING:',
+    'if TYPE_CHECKING:',
+  ]
   run.source          = ["unxt"]
   run.branch          = true
 
@@ -189,6 +193,7 @@
       "PLR2004", # Magic value used in comparison
       "PYI041",  # Use `complex` instead of `int | complex` <- plum is more strict
       "RUF022",  # `__all__` is not sorted
+      "TC003",   # typing-only-standard-library-import
       "TD002",   # Missing author in TODO
       "TD003",   # Missing issue link on the line following this TODO
     ]
