### 2024-12-17 5ba3503
Backport PR #332: 🔧 config(ruff): ignore TC003 and associated fixes (#333)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
