### 2024-12-17 e4579a4
Backport PR #334: 🔧 config(ci): set beartyping in CI
