### 2024-12-17 99c7db5
🔧 config(ci): enable full beartyping in CI (#334)
