+++ src/unxt/_src/units/__init__.py
@@ -5,4 +5,5 @@ Copyright (c) 2023 Galactic Dynamics. All rights reserved.
 
 __all__ = ["unit", "unit_of"]
 
-from .core import unit, unit_of
+from .api import unit, unit_of
+from .core import Unit  # noqa: F401
