+++ src/unxt/_src/quantity/functional.py
@@ -12,7 +12,7 @@ from plum import dispatch
 from .base import AbstractQuantity
 from unxt._src.dimensions import dimension_of
 from unxt._src.units import unit
-from unxt._src.units.core import AbstractUnits
+from unxt._src.units.api import AbstractUnits
 from unxt._src.units.system import AbstractUnitSystem
 
 # ===================================================================
