### 2024-12-17 c3c1fb0
🎨 style(extern): use api more fully
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
