### 2024-12-17 d75d978
✨ feat(quantity): single-argument ustrip
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
