+++ src/unxt/_src/experimental.py
@@ -81,7 +81,7 @@ def grad(
             (a if unit is None else Quantity(a, unit))
             for a, unit in zip(args, theunits, strict=True)
         )
-        return fun(*args_).value  # type: ignore[call-arg]
+        return ustrip(fun(*args_))  # type: ignore[call-arg]
 
     def gradfun(*args: P.args, **kw: P.kwargs) -> R:
         # Get the value of the args. They are turned back into Quantity
@@ -157,7 +157,7 @@ def jacfwd(
         # Adjust the Quantity by the units of the derivative
         # TODO: check the unit correction
         # TODO: get Quantity[unit] / unit2 -> Quantity[unit/unit2] working
-        return type_unparametrized(value)(value.value, value.unit / theunits[argnums])
+        return type_unparametrized(value)(ustrip(value), value.unit / theunits[argnums])
 
     return jacfun
 
@@ -214,7 +214,7 @@ def hessian(
         # TODO: check the unit correction
         # TODO: get Quantity[unit] / unit2 -> Quantity[unit/unit2] working
         return type_unparametrized(value)(
-            value.value, value.unit / theunits[argnums] ** 2
+            ustrip(value), value.unit / theunits[argnums] ** 2
         )
 
     return hessfun
