+++ src/unxt/_src/quantity/register_primitives.py
@@ -62,7 +62,7 @@ def _abs_p(x: AbstractQuantity) -> AbstractQuantity:
     Examples
     --------
     >>> import quaxed.numpy as jnp
-    >>> from unxt import Quantity
+    >>> from unxt.quantity import Quantity
     >>> q = Quantity(-1, "m")
     >>> jnp.abs(q)
     Quantity['length'](Array(1, dtype=int32, ...), unit='m')
@@ -77,7 +77,7 @@ def _abs_p(x: AbstractQuantity) -> AbstractQuantity:
     UncheckedQuantity(Array(1, dtype=int32, ...), unit='m')
 
     """
-    return replace(x, value=lax.abs(x.value))
+    return replace(x, value=lax.abs(ustrip(x)))
 
 
 # ==============================================================================
@@ -95,7 +95,7 @@ def _acos_p_aq(x: AbstractQuantity) -> AbstractQuantity:
     >>> jnp.acos(q)
     UncheckedQuantity(Array(3.1415927, dtype=float32, ...), unit='rad')
 
-    >>> from unxt import Quantity
+    >>> from unxt.quantity import Quantity
     >>> q = Quantity(-1, "")
     >>> jnp.acos(q)
     Quantity['angle'](Array(3.1415927, dtype=float32, ...), unit='rad')
@@ -120,7 +120,7 @@ def _acosh_p_aq(x: AbstractQuantity) -> AbstractQuantity:
     >>> jnp.acosh(q)
     UncheckedQuantity(Array(1.316958, dtype=float32, ...), unit='rad')
 
-    >>> from unxt import Quantity
+    >>> from unxt.quantity import Quantity
     >>> q = Quantity(2.0, "")
     >>> jnp.acosh(q)
     Quantity['angle'](Array(1.316958, dtype=float32, ...), unit='rad')
@@ -150,7 +150,7 @@ def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     >>> q1 + q2
     UncheckedQuantity(Array(1.5, dtype=float32, ...), unit='km')
 
-    >>> from unxt import Quantity
+    >>> from unxt.quantity import Quantity
     >>> q1 = Quantity(1.0, "km")
     >>> q2 = Quantity(500.0, "m")
     >>> jnp.add(q1, q2)
@@ -159,7 +159,7 @@ def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     Quantity['length'](Array(1.5, dtype=float32, ...), unit='km')
 
     """
-    return replace(x, value=lax.add(x.value, ustrip(x.unit, y)))
+    return replace(x, value=lax.add(ustrip(x), ustrip(x.unit, y)))
 
 
 @register(lax.add_p)
@@ -203,7 +203,7 @@ def _add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
 
     `unxt.Quantity`:
 
-    >>> from unxt import Quantity
+    >>> from unxt.quantity import Quantity
     >>> x = jnp.asarray(500.0)
     >>> q2 = Quantity(1.0, "km")
     >>> try:
@@ -226,7 +226,7 @@ def _add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
 
     """
     y = uconvert(one, y)
-    return replace(y, value=lax.add(x, y.value))
+    return replace(y, value=lax.add(x, ustrip(y)))
 
 
 @register(lax.add_p)
[…]
