+++ src/unxt/_src/quantity/base.py
@@ -61,44 +61,44 @@ class AbstractQuantity(
 
     Examples
     --------
-    >>> from unxt import Quantity
+    >>> import unxt as u
 
     From an integer:
 
-    >>> Quantity(1, "m")
+    >>> u.Quantity(1, "m")
     Quantity['length'](Array(1, dtype=int32, ...), unit='m')
 
     From a float:
 
-    >>> Quantity(1.0, "m")
+    >>> u.Quantity(1.0, "m")
     Quantity['length'](Array(1., dtype=float32, ...), unit='m')
 
     From a list:
 
-    >>> Quantity([1, 2, 3], "m")
+    >>> u.Quantity([1, 2, 3], "m")
     Quantity['length'](Array([1, 2, 3], dtype=int32), unit='m')
 
     From a tuple:
 
-    >>> Quantity((1, 2, 3), "m")
+    >>> u.Quantity((1, 2, 3), "m")
     Quantity['length'](Array([1, 2, 3], dtype=int32), unit='m')
 
     From a `numpy.ndarray`:
 
     >>> import numpy as np
-    >>> Quantity(np.array([1, 2, 3]), "m")
+    >>> u.Quantity(np.array([1, 2, 3]), "m")
     Quantity['length'](Array([1, 2, 3], dtype=int32), unit='m')
 
     From a `jax.Array`:
 
     >>> import jax.numpy as jnp
-    >>> Quantity(jnp.array([1, 2, 3]), "m")
+    >>> u.Quantity(jnp.array([1, 2, 3]), "m")
     Quantity['length'](Array([1, 2, 3], dtype=int32), unit='m')
 
     The unit can also be given as a `astropy.units.Unit`:
 
     >>> import astropy.units as apyu
-    >>> Quantity(1, apyu.m)
+    >>> u.Quantity(1, apyu.m)
     Quantity['length'](Array(1, dtype=int32, ...), unit='m')
 
     """
@@ -134,16 +134,16 @@ class AbstractQuantity(
         any subclass of `AbstractQuantity`.
 
         >>> import jax.numpy as jnp
-        >>> from unxt import Quantity
+        >>> import unxt as u
 
         >>> x = jnp.array([1.0, 2, 3])
-        >>> Quantity.from_(x, "m")
+        >>> u.Quantity.from_(x, "m")
         Quantity['length'](Array([1., 2., 3.], dtype=float32), unit='m')
 
-        >>> Quantity.from_([1.0, 2, 3], "m")
+        >>> u.Quantity.from_([1.0, 2, 3], "m")
         Quantity['length'](Array([1., 2., 3.], dtype=float32), unit='m')
 
-        >>> Quantity.from_((1.0, 2, 3), "m")
+        >>> u.Quantity.from_((1.0, 2, 3), "m")
         Quantity['length'](Array([1., 2., 3.], dtype=float32), unit='m')
 
         """
@@ -168,8 +168,8 @@ class AbstractQuantity(
         For this example we'll use the `unxt.Quantity` class. The same applies
         to any subclass of `unxt.AbstractQuantity`.
 
-        >>> from unxt import Quantity
-        >>> Quantity.from_([1.0, 2, 3], unit="m")
+        >>> import unxt as u
+        >>> u.Quantity.from_([1.0, 2, 3], unit="m")
         Quantity['length'](Array([1., 2., 3.], dtype=float32), unit='m')
 
         """
@@ -188,8 +188,8 @@ class AbstractQuantity(
         For this example we'll use the `Quantity` class. The same applies to
         any subclass of `AbstractQuantity`.
 
-        >>> from unxt import Quantity
[…]
