### 2024-12-17 bb82b80
Backport PR #338: ✨ feat(quantity): single-argument ustrip
