### 2024-12-17 3ae70c9
✨ feat(quantity): single-argument ustrip (#338)
* 🎨 style(units): import from api
* ✨ feat(quantity): single-argument ustrip
* 🎨 style(extern): use api more fully
* 🔧 config(pylint): ignore no-value-for-parameter
* 📝 docs(quantity): doctests for abstractquantity

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
