+++ src/unxt/_src/quantity/api.py
@@ -29,7 +29,7 @@ def uconvert(u: Any, x: Any, /) -> Any:
 
 
 @dispatch.abstract  # type: ignore[misc]
-def ustrip(u: Any, x: Any, /) -> Any:
+def ustrip(*args: Any) -> Any:
     """Strip the units from the quantity, first converting if necessary.
 
     Examples
