### 2024-12-17 af9f793
Backport PR #338: ✨ feat(quantity): single-argument ustrip (#339)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
