### 2024-12-17 07f2194
Backport PR #340: ♻️ refactor(quantity): redispatch through primitives
