### 2024-12-17 4ea4978
Backport PR #340: ♻️ refactor(quantity): redispatch through primitives (#341)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
