+++ src/unxt/_src/quantity/register_primitives.py
@@ -77,7 +77,7 @@ def _abs_p(x: AbstractQuantity) -> AbstractQuantity:
     UncheckedQuantity(Array(1, dtype=int32, ...), unit='m')
 
     """
-    return replace(x, value=lax.abs(ustrip(x)))
+    return replace(x, value=qlax.abs(ustrip(x)))
 
 
 # ==============================================================================
@@ -102,7 +102,7 @@ def _acos_p_aq(x: AbstractQuantity) -> AbstractQuantity:
 
     """
     x_ = ustrip(one, x)
-    return type_np(x)(value=lax.acos(x_), unit=radian)
+    return type_np(x)(value=qlax.acos(x_), unit=radian)
 
 
 # ==============================================================================
@@ -127,7 +127,7 @@ def _acosh_p_aq(x: AbstractQuantity) -> AbstractQuantity:
 
     """
     x_ = ustrip(one, x)
-    return type_np(x)(value=lax.acosh(x_), unit=radian)
+    return type_np(x)(value=qlax.acosh(x_), unit=radian)
 
 
 # ==============================================================================
@@ -159,7 +159,7 @@ def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     Quantity['length'](Array(1.5, dtype=float32, ...), unit='km')
 
     """
-    return replace(x, value=lax.add(ustrip(x), ustrip(x.unit, y)))
+    return replace(x, value=qlax.add(ustrip(x), ustrip(x.unit, y)))
 
 
 @register(lax.add_p)
@@ -226,7 +226,7 @@ def _add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
 
     """
     y = uconvert(one, y)
-    return replace(y, value=lax.add(x, ustrip(y)))
+    return replace(y, value=qlax.add(x, ustrip(y)))
 
 
 @register(lax.add_p)
@@ -299,7 +299,7 @@ def _add_p_aqv(x: AbstractQuantity, y: ArrayLike) -> AbstractQuantity:
 
     """
     x = uconvert(one, x)
-    return replace(x, value=lax.add(ustrip(x), y))
+    return replace(x, value=qlax.add(ustrip(x), y))
 
 
 # ==============================================================================
@@ -380,7 +380,7 @@ def _argmax_p(
     UncheckedQuantity(Array(2, dtype=int32), unit='m')
 
     """
-    return replace(operand, value=lax.argmax(ustrip(operand), axes[0], index_dtype))
+    return replace(operand, value=qlax.argmax(ustrip(operand), axes[0], index_dtype))
 
 
 # ==============================================================================
@@ -406,7 +406,7 @@ def _argmin_p(
     UncheckedQuantity(Array(0, dtype=int32), unit='m')
 
     """
-    return replace(operand, value=lax.argmin(ustrip(operand), axes[0], index_dtype))
+    return replace(operand, value=qlax.argmin(ustrip(operand), axes[0], index_dtype))
 
 
 # ==============================================================================
@@ -688,7 +688,7 @@ def _atanh_p_q(
 @register(lax.broadcast_in_dim_p)
 def _broadcast_in_dim_p(operand: AbstractQuantity, **kwargs: Any) -> AbstractQuantity:
     """Broadcast a quantity in a specific dimension."""
-    return replace(operand, value=lax.broadcast_in_dim(ustrip(operand), **kwargs))
+    return replace(operand, value=qlax.broadcast_in_dim(ustrip(operand), **kwargs))
 
 
 # ==============================================================================
[…]
