+++ src/unxt/_src/quantity/register_primitives.py
@@ -30,6 +30,7 @@ from .base import AbstractQuantity
 from .base_parametric import AbstractParametricQuantity
 from .core import Quantity
 from unxt._src.units import unit, unit_of
+from unxt._src.utils import promote_dtypes
 
 T = TypeVar("T")
 
@@ -143,7 +144,7 @@ def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     >>> import quaxed.numpy as jnp
     >>> from unxt.quantity import UncheckedQuantity
 
-    >>> q1 = UncheckedQuantity(1.0, "km")
+    >>> q1 = UncheckedQuantity(1, "km")
     >>> q2 = UncheckedQuantity(500.0, "m")
     >>> jnp.add(q1, q2)
     UncheckedQuantity(Array(1.5, dtype=float32, ...), unit='km')
@@ -151,7 +152,7 @@ def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     UncheckedQuantity(Array(1.5, dtype=float32, ...), unit='km')
 
     >>> from unxt.quantity import Quantity
-    >>> q1 = Quantity(1.0, "km")
+    >>> q1 = Quantity(1, "km")
     >>> q2 = Quantity(500.0, "m")
     >>> jnp.add(q1, q2)
     Quantity['length'](Array(1.5, dtype=float32, ...), unit='km')
@@ -170,7 +171,7 @@ def _add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
     --------
     >>> import quaxed.numpy as jnp
 
-    >>> x = jnp.asarray(500.0)
+    >>> x = jnp.asarray(500)
 
     `unxt.UncheckedQuantity`:
 
@@ -237,7 +238,7 @@ def _add_p_aqv(x: AbstractQuantity, y: ArrayLike) -> AbstractQuantity:
     --------
     >>> import quaxed.numpy as jnp
 
-    >>> y = jnp.asarray(500.0)
+    >>> y = jnp.asarray(500)
 
     `unxt.UncheckedQuantity`:
 
@@ -317,7 +318,7 @@ def _add_any_p(
     >>> import quaxed.numpy as jnp
     >>> import unxt as u
 
-    >>> q1 = u.Quantity(1.0, "km")
+    >>> q1 = u.Quantity(1, "km")
     >>> q2 = u.Quantity(500.0, "m")
 
     >>> @jax.jit
@@ -495,14 +496,14 @@ def _atan2_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     >>> import quaxed.numpy as jnp
     >>> from unxt.quantity import UncheckedQuantity
     >>> q1 = UncheckedQuantity(1, "m")
-    >>> q2 = UncheckedQuantity(3, "m")
+    >>> q2 = UncheckedQuantity(3.0, "m")
     >>> jnp.atan2(q1, q2)
     UncheckedQuantity(Array(0.32175055, dtype=float32, ...), unit='rad')
 
     """
     x, y = promote(x, y)  # e.g. Distance -> Quantity
-    y_ = ustrip(x.unit, y)
-    return type_np(x)(lax.atan2(ustrip(x), y_), unit=radian)
+    yv = ustrip(x.unit, y)
+    return type_np(x)(lax.atan2(ustrip(x), yv), unit=radian)
 
 
 @register(lax.atan2_p)
@@ -516,14 +517,14 @@ def _atan2_p_qq(
     >>> import quaxed.numpy as jnp
[…]
