### 2024-12-19 8914e70
♻️ refactor(quantity): more adjustments to the dtype promotion
