### 2024-12-20 82e210b
Backport PR #346: 🐛 fix(quantity): fix gather_p
