### 2024-12-22 791e48e
Backport PR #348: ♻️ refactor(quantity): ensure abstract quantities are abstract
