### 2024-12-22 5bb134d
♻️ refactor(quantity): ensure abstract quantities are abstract (#348)
