+++ src/unxt/_src/quantity/fast.py
@@ -5,7 +5,13 @@ __all__ = ["UncheckedQuantity"]
 
 from typing import Any
 
+import equinox as eqx
+import jax
+from jaxtyping import Array, Shaped
+
 from .base import AbstractQuantity
+from unxt._src.units import unit as parse_unit
+from unxt._src.units.api import AbstractUnits
 
 
 class UncheckedQuantity(AbstractQuantity):
@@ -13,6 +19,12 @@ class UncheckedQuantity(AbstractQuantity):
 
     This class is not parametrized by its dimensionality.
     """
+
+    value: Shaped[Array, "*shape"] = eqx.field(converter=jax.numpy.asarray)
+    """The value of the `AbstractQuantity`."""
+
+    unit: AbstractUnits = eqx.field(static=True, converter=parse_unit)
+    """The unit associated with this value."""
 
     def __class_getitem__(
         cls: type["UncheckedQuantity"], item: Any
