### 2024-12-22 e069047
🎨 style(quantity): better layout for constructors
