### 2024-12-22 167dff9
Backport PR #348: ♻️ refactor(quantity): ensure abstract quantities are abstract (#349)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
