### 2024-12-23 52456b6
✨ feat(quantity): add a function is_any_quantity checking isinstance(x, AbstractQuantity)
This is useful for type checking, and also for jax.tree is_leaf filtering.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
