### 2024-12-23 27aa9df
Backport PR #352: 🎨 style(quantity): rename modules
