### 2025-01-06 deadc66
➕ dep-add(extras): add extra dependencies
enable mutually exclusive extra dependencies

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
