### 2025-01-07 51a6d0a
✨ feat(quantity): enable non-Array quax.ArrayValue as Quantity's value
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
