+++ src/unxt/_src/quantity/__init__.py
@@ -9,6 +9,7 @@ __all__ = [
     "uconvert",
     "ustrip",
     "is_any_quantity",
+    "value_converter",
 ]
 
 from .api import is_unit_convertible, uconvert, ustrip
@@ -16,3 +17,4 @@ from .base import AbstractQuantity, is_any_quantity
 from .base_parametric import AbstractParametricQuantity
 from .quantity import Quantity
 from .unchecked import UncheckedQuantity
+from .value import value_converter
