### 2025-01-07 d4a6704
⬆️ dep-bump(plum): bump v2.5.5+ for more type info
