### 2025-01-07 ccf8fda
👷 build(mypy): add plum to mypy in pre-commit
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
