+++ src/unxt/_src/units/system/compare.py
@@ -7,7 +7,7 @@ from plum import dispatch
 from .base import AbstractUnitSystem
 
 
-@dispatch  # type: ignore[misc]
+@dispatch
 def equivalent(a: AbstractUnitSystem, b: AbstractUnitSystem, /) -> bool:
     """Check if two unit systems are equivalent."""
     return set(a.base_dimensions) == set(b.base_dimensions)
