### 2025-01-07 2a1a07b
Backport PR #359: ⬆️ dep-bump(plum): bump v2.5.5+ for more type info
