### 2025-01-07 c588f88
⬆️ dep-bump(plum): bump v2.5.5+ for more type info (#359)
* ⬆️ dep-bump(plum): bump v2.5.5+ for more type info
* 👷 build(mypy): add plum to mypy in pre-commit

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
