### 2025-01-08 5e85c6a
Backport PR #359: ⬆️ dep-bump(plum): bump v2.5.5+ for more type info (#360)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
