### 2025-01-08 e71bbf2
✅ test(benchmarks): add benchmark test for dimensions
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
