### 2025-01-08 aefd9d4
TST: benchmarks (#361)
* ✅ test(benchmarks): add benchmark test for dimensions
* ✅ test(benchmarks): units benchmarks
* 👷 build(benchmarks): run benchmarks on mergest to main

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
