### 2025-01-08 89da29b
👷 build(benchmarks): add more benchmarks and config
