### 2025-01-08 b89ded5
Backport PR #361: TST: benchmarks (#362)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
