### 2025-01-08 9d3d51f
⬆️ dep-bump(deps): bump and set optional dependency pins
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
