+++ pyproject.toml
@@ -54,7 +54,7 @@
     all             = ["unxt[backend-astropy,interop-gala,interop-mpl]"]
     backend-astropy = ["astropy>=6.0"]
     interop-gala    = ["gala>=1.8"]
-    interop-mpl     = ["matplotlib>=3.4"]
+    interop-mpl     = ["matplotlib>=3.5"]
 
   [project.urls]
     "Bug Tracker" = "https://github.com/GalacticDynamics/unxt/issues"
@@ -80,12 +80,12 @@
     "pytz>=2024.2",               # for copyright date
     "sphinx-book-theme==1.1.3",
     "sphinx-prompt>=1.8.0",
-    "sphinx-tippy",
+    "sphinx-tippy>=0.4.3",
     "sphinx>=7.0",
-    "sphinx_autodoc_typehints",
-    "sphinx_copybutton",
-    "sphinx_design",
-    "sphinx_togglebutton",
+    "sphinx_autodoc_typehints>=3.0.0",
+    "sphinx_copybutton>=0.5.2",
+    "sphinx_design>=0.6.1",
+    "sphinx_togglebutton>=0.3.2",
     "sphinxext-opengraph>=0.9.1",
     "sphinxext-rediraffe>=0.2.7",
   ]
@@ -104,7 +104,7 @@
   pytest-benchmark-parallel = [
     "pytest-xdist>=3.6.1",
   ]
-  test-mpl = ["pytest-mpl"]
+  test-mpl = ["pytest-mpl>=0.17.0"]
   test-all = [{ include-group = "test" }, { include-group = "test-mpl" }]
 
 
