### 2025-01-08 a9e0afb
👷 build(benchmarks): add more benchmarks and config (#363)
* 👷 build(benchmarks): add more benchmarks and config
* ⬆️ dep-bump(deps): bump and set optional dependency pins

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
