### 2025-01-08 bf14c55
Backport PR #363: 👷 build(benchmarks): add more benchmarks and config
