### 2025-01-08 df642c2
Backport PR #363: 👷 build(benchmarks): add more benchmarks and config (#364)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
