### 2025-01-14 #7583 @danielskatz
👋 @nstarman - thanks for your submission. I'll handle it as editor.

1. Please add a references header at the end of your .md file. Please see the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). After you make changes to your .md file, use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

2. Please suggest ~5 potential reviewers. You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Or people who aren't in the JOSS system would also be ok.  If you know their GitHub usernames, then please mention them here in this thread (without tagging them with an @).
