### 2025-01-15 91a43b7
💚 ci(mypy): pin plum for pre-commit
