+++ .pre-commit-config.yaml
@@ -79,7 +79,7 @@ repos:
       - id: mypy
         files: src
         additional_dependencies:
-          - plum-dispatch
+          - plum-dispatch>=2.5.6
 
   - repo: https://github.com/codespell-project/codespell
     rev: "v2.3.0"
