+++ src/unxt/unitsystems.py
@@ -46,7 +46,7 @@ from jaxtyping import install_import_hook
 from .setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("unxt.unitsystems", RUNTIME_TYPECHECKER):
-    from ._src.units.system import (
+    from ._src.unitsystems import (
         NAMED_UNIT_SYSTEMS,
         UNITSYSTEMS_REGISTRY,
         AbstractUnitSystem,
