+++ src/unxt/_src/quantity/register_api.py
@@ -14,7 +14,7 @@ from .base import AbstractQuantity
 from unxt._src.dimensions import AbstractDimension, dimension_of
 from unxt._src.units import unit
 from unxt._src.units.api import AbstractUnits
-from unxt._src.units.system import AbstractUnitSystem
+from unxt._src.unitsystems import AbstractUnitSystem
 
 # ===================================================================
 # Get dimensions
