### 2025-01-20 #7583 @danielskatz
👋 @ngoldbaum & @Cadair - would one or both of you be willing to review this JOSS submission?
