### 2025-01-20 #7583 @danielskatz
👋 @PhilipVinc,@yjdeveloper, @thomaspinder - Would one or two of you be willing to review this JOSS submission?
