### 2025-01-20 #7583 @ngoldbaum
I don't have time to review this. But given the naming similarities with unyt maybe you could cite the unyt JOSS paper? Even if they're unrelated it still seems worth citing as prior art for the usefulness of a unit library.

You might also be interested in the array-api compatible quantities discussions happening over at https://github.com/orgs/quantity-dev/discussions.
