### 2025-01-20 0ff6e44
📝 docs(joss): add section on related works
