### 2025-01-20 3c2b9b6
📝 docs(joss): add section on related works
