+++ paper/paper.bib
@@ -1,3 +1,10 @@
+@misc{array-api,
+  author = {Array API Consortion},
+  title  = {Array API Standard},
+  url    = {https://data-apis.org/array-api/latest/index.html},
+  year   = {2021}
+}
+
 @article{astropy:13,
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
   adsurl        = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..33A},
@@ -35,6 +42,13 @@
   year          = 2022
 }
 
+@manual{dask:2016,
+  author = {{Dask Development Team}},
+  title  = {Dask: Library for dynamic task scheduling},
+  url    = {http://dask.pydata.org},
+  year   = {2016}
+}
+
 @software{jax:18,
   author  = {Bradbury, James and Frostig, Roy and Hawkins, Peter and Johnson, Matthew James and Leary, Chris and Maclaurin, Dougal and Necula, George and Paszke, Adam and VanderPlas, Jake and Wanderman-Milne, Skye and Zhang, Qiao},
   date    = {2018},
@@ -42,6 +56,7 @@
   url     = {http://github.com/jax-ml/jax},
   version = {0.4.35}
 }
+
 
 @misc{nasa:98,
   author       = {NASA},
@@ -51,9 +66,52 @@
   year         = {}
 }
 
+@article{numpy:2020,
+  author    = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+               van der Walt and Ralf Gommers and Pauli Virtanen and David
+               Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+               Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+               and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+               Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+               R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+               G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+               Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+               Travis E. Oliphant},
+  doi       = {10.1038/s41586-020-2649-2},
+  journal   = {Nature},
+  month     = sep,
+  number    = {7825},
+  pages     = {357--362},
+  publisher = {Springer Science and Business Media {LLC}},
+  title     = {Array programming with {NumPy}},
+  url       = {https://doi.org/10.1038/s41586-020-2649-2},
+  volume    = {585},
+  year      = {2020}
+}
+
+@software{pint,
+  author = {Grecco, Hernan},
+  date   = {2012},
+  title  = {Pint: makes units easy}
+}
+
 @software{quax:23,
   author = {Kidger, Patrrick},
   date   = {2023},
   title  = {{{Quax}}: {{JAX}} + multiple dispatch + custom array-ish objects},
   year   = {}
+}
+
+@article{unyt:2018,
+  author    = {Nathan J. Goldbaum and John A. ZuHone and Matthew J. Turk and Kacper Kowalik and Anna L. Rosen},
+  doi       = {10.21105/joss.00809},
+  journal   = {Journal of Open Source Software},
+  month     = {aug},
+  number    = {28},
+  pages     = {809},
+  publisher = {The Open Journal},
+  title     = {unyt: Handle,  manipulate,  and convert data with units in Python},
+  url       = {https://doi.org/10.21105/joss.00809},
+  volume    = {3},
+  year      = {2018}
 }
