### 2025-01-20 #7583 @nstarman
@danielskatz, does it mess anything up for me to remove old `(auto) Paper PDF` commits from my [branch's](https://github.com/GalacticDynamics/unxt/tree/joss) git history and force-push?
