### 2025-01-20 #7583 @danielskatz
> [@danielskatz](https://github.com/danielskatz), does it mess anything up for me to remove old `(auto) Paper PDF` commits from my [branch's](https://github.com/GalacticDynamics/unxt/tree/joss) [git history](https://github.com/GalacticDynamics/unxt/compare/main...joss) and force-push?

No, JOSS doesn't need the pdf of the paper in the repo at all, as we build it from source here. And you can make whatever changes in the repo you want, at least until we start the review, then changes related to the code and paper in the review should be communicated to the reviewers
