### 2025-01-21 23e39a2
👽️ external(quaxed): use new quaxed.experimental.arrayish module
