+++ src/unxt/_src/quantity/register_primitives.py
@@ -3,7 +3,7 @@
 
 __all__: list[str] = []
 
-from collections.abc import Callable, Sequence
+from collections.abc import Sequence
 from dataclasses import replace
 from math import prod
 from typing import Any, TypeAlias, TypeVar
@@ -17,11 +17,10 @@ from astropy.units import (  # pylint: disable=no-name-in-module
 )
 from jax import lax, numpy as jnp
 from jax._src.ad_util import add_any_p
-from jax.core import Primitive
 from jaxtyping import Array, ArrayLike
 from plum import promote
 from plum.parametric import type_unparametrized as type_np
-from quax import register as register_
+from quax import register
 
 from quaxed import lax as qlax
 
@@ -35,11 +34,6 @@ from unxt._src.utils import promote_dtypes_if_needed
 T = TypeVar("T")
 
 Axes: TypeAlias = tuple[int, ...]
-
-
-def register(primitive: Primitive, **kwargs: Any) -> Callable[[T], T]:
-    """`quax.register`, but makes mypy happy."""
-    return register_(primitive, **kwargs)
 
 
 def _to_value_rad_or_one(q: AbstractQuantity) -> ArrayLike:
@@ -334,7 +328,7 @@ def _add_any_p(
     Quantity['length'](Array(1.5, dtype=float32, ...), unit='km')
 
     """
-    return replace(x, value=add_any_p.bind(ustrip(x), ustrip(y)))
+    return replace(x, value=add_any_p.bind(ustrip(x), ustrip(y)))  # type: ignore[no-untyped-call]
 
 
 # ==============================================================================
@@ -1078,9 +1072,7 @@ def _cond_p_vq(
     >>> from unxt import Quantity
 
     """
-    # print(branches)
-    # raise AttributeError
-    return lax.cond_p.bind(index, ustrip(consts), branches=branches)
+    return lax.cond_p.bind(index, ustrip(consts), branches=branches)  # type: ignore[no-untyped-call]
 
 
 # ==============================================================================
@@ -1119,7 +1111,8 @@ def _convert_element_type_p(
     """Convert the element type of a quantity."""
     # TODO: examples
     return replace(
-        operand, value=lax.convert_element_type_p.bind(ustrip(operand), **kwargs)
+        operand,
+        value=lax.convert_element_type_p.bind(ustrip(operand), **kwargs),  # type: ignore[no-untyped-call]
     )
 
 
@@ -1146,7 +1139,7 @@ def _copy_p(x: AbstractQuantity) -> AbstractQuantity:
     Quantity['length'](Array(1, dtype=int32, ...), unit='m')
 
     """
-    return replace(x, value=lax.copy_p.bind(ustrip(x)))
+    return replace(x, value=lax.copy_p.bind(ustrip(x)))  # type: ignore[no-untyped-call]
 
 
 # ==============================================================================
@@ -1408,7 +1401,7 @@ def _device_put_p(x: AbstractQuantity, **kwargs: Any) -> AbstractQuantity:
     Quantity['length'](Array(1, dtype=int32, ...), unit='m')
 
     """
-    return jt.map(lambda y: lax.device_put_p.bind(y, **kwargs), x)
+    return jt.map(lambda y: lax.device_put_p.bind(y, **kwargs), x)  # type: ignore[no-untyped-call]
 
 
 # ==============================================================================
@@ -1865,7 +1858,7 @@ def _exp2_p(x: AbstractQuantity) -> AbstractQuantity:
     Quantity['dimensionless'](Array(8., dtype=float32, ...), unit='')
 
     """
-    return replace(x, value=qlax.exp2(ustrip(one, x)))
[…]
