### 2025-01-21 1090114
💚 ci(pre-commit): update to manual pre-commit
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
