+++ .github/workflows/ci.yml
@@ -37,9 +37,8 @@ jobs:
       - name: Install the project
         run: uv sync --group nox
 
-      #     - uses: pre-commit/action@v3.0.1
-      #       with:
-      #         extra_args: --hook-stage manual --all-files
+      - uses: pre-commit/action@v3.0.1
+
       - name: Run PyLint
         run: uv run nox -s pylint -- --output-format=github
 
