### 2025-01-21 66501e9
🎨 style(dimensions): restructure private files
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
