### 2025-01-21 833dc6e
Backport PR #372: 🎨 style(dimensions): restructure private files
