+++ src/unxt/_src/dimensions/__init__.py
@@ -6,4 +6,8 @@ This is the private implementation of the dimensions module.
 
 __all__ = ["AbstractDimension", "dimension", "dimension_of"]
 
-from .core import AbstractDimension, dimension, dimension_of
+from .api import AbstractDimension, dimension, dimension_of
+
+# Register the dispatches
+# isort: split
+from . import core  # noqa: F401
