### 2025-01-21 04e04f0
Backport PR #372: 🎨 style(dimensions): restructure private files (#373)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
