### 2025-01-22 49de864
✨ feat(quantity): add a function is_any_quantity checking isinstance(x, AbstractQuantity) (#354)
This is useful for type checking, and also for jax.tree is_leaf filtering.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
