+++ src/unxt/_src/quantity/__init__.py
@@ -8,10 +8,11 @@ __all__ = [
     "is_unit_convertible",
     "uconvert",
     "ustrip",
+    "is_any_quantity",
 ]
 
 from .api import is_unit_convertible, uconvert, ustrip
-from .base import AbstractQuantity
+from .base import AbstractQuantity, is_any_quantity
 from .base_parametric import AbstractParametricQuantity
 from .quantity import Quantity
 from .unchecked import UncheckedQuantity
