### 2025-01-22 c7ae4f5
👽️ external(quaxed): use new quaxed.experimental.arrayish module (#370)
* 👽️ external(quaxed): use new quaxed.experimental.arrayish module
* 💚 ci(pre-commit): update to manual pre-commit

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
