+++ src/unxt/_src/quantity/base.py
@@ -7,19 +7,20 @@ from collections.abc import Callable, Mapping
 from functools import partial
 from types import ModuleType
 from typing import TYPE_CHECKING, Any, ClassVar, NoReturn, TypeAlias, TypeGuard, TypeVar
+from typing_extensions import override
 
 import equinox as eqx
 import jax
 import jax.core
 from astropy.units import UnitConversionError
 from jax._src.numpy.array_methods import _IndexUpdateHelper, _IndexUpdateRef
-from jaxtyping import Array, ArrayLike, ScalarLike, Shaped
+from jaxtyping import Array, ArrayLike, Bool, ScalarLike, Shaped
 from plum import add_promotion_rule, dispatch
 from quax import ArrayValue
 
 import quaxed.numpy as jnp
-import quaxed.operator as qoperator
-from dataclassish import fields, replace
+from dataclassish import replace
+from quaxed.experimental import arrayish
 
 from .api import is_unit_convertible, uconvert, ustrip
 from .mixins import AstropyQuantityCompatMixin, IPythonReprMixin, NumPyCompatMixin
@@ -55,7 +56,16 @@ class AbstractQuantity(
     AstropyQuantityCompatMixin,
     NumPyCompatMixin,
     IPythonReprMixin,
-    ArrayValue,  # type: ignore[misc]
+    ArrayValue,
+    arrayish.NumpyBinaryOpsMixin[Any, "AbstractQuantity"],
+    arrayish.NumpyComparisonMixin[Any, Bool[Array, "*shape"]],  # TODO: shape hint
+    arrayish.NumpyUnaryMixin["AbstractQuantity"],
+    arrayish.NumpyRoundMixin["AbstractQuantity"],
+    arrayish.NumpyTruncMixin["AbstractQuantity"],
+    arrayish.NumpyFloorMixin["AbstractQuantity"],
+    arrayish.NumpyCeilMixin["AbstractQuantity"],
+    arrayish.LaxLenMixin,
+    arrayish.LaxLengthHintMixin,
 ):
     """Represents an array, with each axis bound to a name.
 
@@ -175,7 +185,7 @@ class AbstractQuantity(
         raise RuntimeError(msg)
 
     def aval(self) -> jax.core.ShapedArray:
-        return jax.core.get_aval(self.value)
+        return jax.core.get_aval(self.value)  # type: ignore[no-untyped-call]
 
     def enable_materialise(self, _: bool = True) -> "Self":  # noqa: FBT001, FBT002
         return replace(self, value=self.value, unit=self.unit)
@@ -289,47 +299,6 @@ class AbstractQuantity(
     # ---------------------------------------------------------------
     # arithmetic operators
 
-    def __pos__(self) -> "AbstractQuantity":
-        """Return the value of the array.
-
-        Examples
-        --------
-        >>> import unxt as u
-        >>> q = u.Quantity(1, "m")
-        >>> +q
-        Quantity['length'](Array(1, dtype=int32, ...), unit='m')
-
-        """
-        return replace(self, value=+self.value)  # pylint: disable=E1130
-
-    def __neg__(self) -> "AbstractQuantity":
-        """Negate the value of the array.
-
-        Examples
-        --------
-        >>> import unxt as u
-        >>> q = u.Quantity(1, "m")
-        >>> -q
-        Quantity['length'](Array(-1, dtype=int32, ...), unit='m')
-
-        """
-        return replace(self, value=-self.value)  # pylint: disable=E1130
-
-    __add__ = qoperator.add
-    __radd__ = _flip_binop(qoperator.add)
-
-    __sub__ = qoperator.sub
-    __rsub__ = _flip_binop(qoperator.sub)
-
-    __mul__ = qoperator.mul
-    __rmul__ = _flip_binop(qoperator.mul)
-
[…]
