+++ src/unxt/_src/quantity/__init__.py
@@ -9,7 +9,7 @@ __all__ = [
     "uconvert",
     "ustrip",
     "is_any_quantity",
-    "value_converter",
+    "convert_to_quantity_value",
 ]
 
 from .api import is_unit_convertible, uconvert, ustrip
@@ -17,4 +17,4 @@ from .base import AbstractQuantity, is_any_quantity
 from .base_parametric import AbstractParametricQuantity
 from .quantity import Quantity
 from .unchecked import UncheckedQuantity
-from .value import value_converter
+from .value import convert_to_quantity_value
