+++ src/unxt/_src/quantity/quantity.py
@@ -12,7 +12,7 @@ from plum import parametric
 
 from .base import AbstractQuantity
 from .base_parametric import AbstractParametricQuantity
-from .value import value_converter
+from .value import convert_to_quantity_value
 from unxt._src.units import AstropyUnits, unit as parse_unit
 from unxt.units import unit as parse_unit
 
@@ -106,7 +106,7 @@ class Quantity(AbstractParametricQuantity):
 
     """
 
-    value: Shaped[Array, "*shape"] = eqx.field(converter=value_converter)
+    value: Shaped[Array, "*shape"] = eqx.field(converter=convert_to_quantity_value)
     """The value of the `AbstractQuantity`."""
 
     unit: AstropyUnits = eqx.field(static=True, converter=parse_unit)
