+++ src/unxt/_src/quantity/unchecked.py
@@ -9,7 +9,7 @@ import equinox as eqx
 from jaxtyping import Array, Shaped
 
 from .base import AbstractQuantity
-from .value import value_converter
+from .value import convert_to_quantity_value
 from unxt._src.units import AstropyUnits, unit as parse_unit
 from unxt.units import unit as parse_unit
 
@@ -20,7 +20,7 @@ class UncheckedQuantity(AbstractQuantity):
     This class is not parametrized by its dimensionality.
     """
 
-    value: Shaped[Array, "*shape"] = eqx.field(converter=value_converter)
+    value: Shaped[Array, "*shape"] = eqx.field(converter=convert_to_quantity_value)
     """The value of the `AbstractQuantity`."""
 
     unit: AstropyUnits = eqx.field(static=True, converter=parse_unit)
