### 2025-01-23 49ddc12
♻️ refactor(units): consolidate abstracts (#374) (#375)
(cherry picked from commit 53cdf12fc5ebaace4e4378a1e224eaf096bd6fc2)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
