### 2025-01-23 ac17545
🔧 config(flake8): equinox import alias
