### 2025-01-23 5aba0b7
🔥 dump(quantity): rm unused utilities (#377)
