### 2025-01-25 b9d67be
♻️ refactor(zeroth): replace next-iter with zeroth
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
