### 2025-01-25 ce25866
♻️ refactor(zeroth): replace next-iter with zeroth (#378)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
