### 2025-01-25 de0feed
🏷️ types(units): consolidate type hints
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
