### 2025-01-25 8b04cd0
🏷️ types(units): consolidate type hints (#379)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
