+++ src/unxt/_src/unitsystems/base.py
@@ -14,8 +14,8 @@ from is_annotated import isannotated
 
 from .utils import get_dimension_name
 from unxt._src.dimensions import AbstractDimension, dimension
-from unxt._src.typing_ext import Unit as UnitT
 from unxt._src.units import unit
+from unxt._src.units.api import AbstractUnits
 
 Unit = AstropyUnitBase
 
@@ -112,14 +112,14 @@ class AbstractUnitSystem:
         return self._base_dimensions
 
     @property
-    def base_units(self) -> tuple[UnitT, ...]:
+    def base_units(self) -> tuple[AbstractUnits, ...]:
         """List of core units."""
         return tuple(getattr(self, k) for k in self._base_field_names)
 
     # ===============================================================
     # Python stuff
 
-    def __getitem__(self, key: AbstractDimension | str) -> UnitT:
+    def __getitem__(self, key: AbstractDimension | str) -> AbstractUnits:
         """Get the unit for a given physical type.
 
         Examples
@@ -142,7 +142,7 @@ class AbstractUnitSystem:
         if key in self.base_dimensions:
             return getattr(self, get_dimension_name(key))
 
-        out: UnitT
+        out: AbstractUnits
         for k, v in _physical_unit_mapping.items():
             if v == key:
                 out = unit(" ".join([f"{x}**{y}" for x, y in k]))
@@ -171,7 +171,7 @@ class AbstractUnitSystem:
         return len(self.base_dimensions)
 
     # TODO: should this be changed to _base_field_names -> Iterator[str]?
-    def __iter__(self) -> Iterator[UnitT]:
+    def __iter__(self) -> Iterator[AbstractUnits]:
         """Iterate over the base units.
 
         This is different than a dictionary, which would iterate over the
