### 2025-01-25 e558b58
🎨 style(quantity): improve primitives (#381)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
