+++ src/unxt/_src/quantity/register_primitives.py
@@ -40,10 +40,6 @@ def _to_value_rad_or_one(q: AbstractQuantity) -> ArrayLike:
     return ustrip(radian if is_unit_convertible(q.unit, radian) else one, q)
 
 
-def _bshape(arrs: tuple[Any, ...], /) -> tuple[int, ...]:
-    return jnp.broadcast_shapes(*map(jnp.shape, arrs))
-
-
 ################################################################################
 # Registering Primitives
 
@@ -51,7 +47,7 @@ def _bshape(arrs: tuple[Any, ...], /) -> tuple[int, ...]:
 
 
 @register(lax.abs_p)
-def _abs_p(x: AbstractQuantity) -> AbstractQuantity:
+def abs_p(x: AbstractQuantity) -> AbstractQuantity:
     """Absolute value of a quantity.
 
     Examples
@@ -79,7 +75,7 @@ def _abs_p(x: AbstractQuantity) -> AbstractQuantity:
 
 
 @register(lax.acos_p)
-def _acos_p_aq(x: AbstractQuantity) -> AbstractQuantity:
+def acos_p_aq(x: AbstractQuantity) -> AbstractQuantity:
     """Inverse cosine of a quantity.
 
     Examples
@@ -104,7 +100,7 @@ def _acos_p_aq(x: AbstractQuantity) -> AbstractQuantity:
 
 
 @register(lax.acosh_p)
-def _acosh_p_aq(x: AbstractQuantity) -> AbstractQuantity:
+def acosh_p_aq(x: AbstractQuantity) -> AbstractQuantity:
     """Inverse hyperbolic cosine of a quantity.
 
     Examples
@@ -130,7 +126,7 @@ def _acosh_p_aq(x: AbstractQuantity) -> AbstractQuantity:
 
 
 @register(lax.add_p)
-def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
+def add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     """Add two quantities.
 
     Examples
@@ -163,7 +159,7 @@ def _add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
 
 
 @register(lax.add_p)
-def _add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
+def add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
     """Add a value and a quantity.
 
     Examples
@@ -230,7 +226,7 @@ def _add_p_vaq(x: ArrayLike, y: AbstractQuantity) -> AbstractQuantity:
 
 
 @register(lax.add_p)
-def _add_p_aqv(x: AbstractQuantity, y: ArrayLike) -> AbstractQuantity:
+def add_p_aqv(x: AbstractQuantity, y: ArrayLike) -> AbstractQuantity:
     """Add a quantity and a value.
 
     Examples
@@ -306,7 +302,7 @@ def _add_p_aqv(x: AbstractQuantity, y: ArrayLike) -> AbstractQuantity:
 
 
 @register(add_any_p)
-def _add_any_p(
+def add_any_p(
     x: AbstractParametricQuantity, y: AbstractParametricQuantity
 ) -> AbstractParametricQuantity:
     """Add two quantities using the ``jax._src.ad_util.add_any_p``.
@@ -335,7 +331,7 @@ def _add_any_p(
 
 
 @register(lax.and_p)
-def _and_p_aq(x1: AbstractQuantity, x2: AbstractQuantity, /) -> ArrayLike:
[…]
