### 2025-01-25 cc730a7
✨ feat(quantity): add promotion rule
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
