### 2025-01-26 1b40931
🎨 style(dims): use public API
