+++ src/unxt/_src/unitsystems/core.py
@@ -19,8 +19,8 @@ from .builtin import DimensionlessUnitSystem
 from .flags import AbstractUSysFlag, DynamicalSimUSysFlag, StandardUSysFlag
 from .realizations import NAMED_UNIT_SYSTEMS, dimensionless
 from .utils import get_dimension_name
-from unxt._src.units import unit
 from unxt.dims import dimension_of
+from unxt.units import unit
 
 # ===================================================================
 # `unitsystem` function
