+++ src/unxt/_src/units/__init__.py
@@ -3,7 +3,11 @@
 Copyright (c) 2023 Galactic Dynamics. All rights reserved.
 """
 
-__all__ = ["unit", "unit_of"]
+__all__ = ["unit", "unit_of", "AbstractUnits"]
 
-from .api import unit, unit_of
-from .core import Unit  # noqa: F401
+from .api import (
+    AbstractUnits,  # TODO: remove
+    unit,
+    unit_of,
+)
+from .core import Unit  # noqa: F401  # TODO: remove
