+++ src/unxt/_src/quantity/register_primitives.py
@@ -28,8 +28,8 @@ from .api import is_unit_convertible, uconvert, ustrip
 from .base import AbstractQuantity
 from .base_parametric import AbstractParametricQuantity
 from .quantity import Quantity
-from unxt._src.units import unit, unit_of
 from unxt._src.utils import promote_dtypes_if_needed
+from unxt.units import unit, unit_of
 
 T = TypeVar("T")
 
