+++ src/unxt/_src/quantity/unchecked.py
@@ -10,8 +10,7 @@ import jax
 from jaxtyping import Array, Shaped
 
 from .base import AbstractQuantity
-from unxt._src.units import unit as parse_unit
-from unxt._src.units.api import AbstractUnits
+from unxt.units import AbstractUnits, unit as parse_unit
 
 
 class UncheckedQuantity(AbstractQuantity):
