+++ src/unxt/_src/quantity/register_api.py
@@ -11,9 +11,9 @@ from plum import dispatch
 
 from .api import ustrip
 from .base import AbstractQuantity
-from unxt._src.unitsystems import AbstractUnitSystem
 from unxt.dims import AbstractDimension, dimension_of
 from unxt.units import AbstractUnits, unit
+from unxt.unitsystems import AbstractUnitSystem
 
 # ===================================================================
 # Get dimensions
