### 2025-01-26 d8ad479
🎨 style(dims): clean up use of public API in private modules (#383)
* 🎨 style(dims): use public API
* 🎨 style(units): use public API
* 🎨 style(usys): use public API

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
