+++ src/unxt/dims.py
@@ -7,14 +7,14 @@ The main features are:
 
 """
 
-__all__ = ["dimension", "dimension_of"]
+__all__ = ["dimension", "dimension_of", "AbstractDimension"]
 
 from jaxtyping import install_import_hook
 
 from .setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("unxt.dims", RUNTIME_TYPECHECKER):
-    from ._src.dimensions import dimension, dimension_of
+    from ._src.dimensions import AbstractDimension, dimension, dimension_of
 
 # Clean up the namespace
 del install_import_hook
