+++ src/unxt/_src/units/register_dispatches.py
@@ -3,16 +3,11 @@
 Copyright (c) 2023 Galactic Dynamics. All rights reserved.
 """
 
-__all__ = [
-    "Unit",  # legacy
-    "unit",
-    "unit_of",
-]
+__all__: list[str] = []
 
 from typing import Any
 
 import astropy.units as apyu
-from astropy.units import Unit
 from plum import dispatch
 
 from .api import AbstractUnits
