### 2025-01-26 07c1d0d
🎨 style(units): rename private Astropy type alias
