+++ src/unxt/_interop/unxt_interop_astropy/custom_types.py
@@ -1,0 +1,12 @@
+"""Custom type annotations for astropy."""
+
+__all__ = ["APYUnits"]
+
+
+from typing import TypeAlias
+
+import astropy.units as apyu
+
+APYUnits: TypeAlias = (
+    apyu.UnitBase | apyu.Unit | apyu.FunctionUnitBase | apyu.StructuredUnit
+)
