### 2025-01-26 81ed4eb
🎨 style(units): rename private Astropy type alias (#385)
