### 2025-01-29 d90350f
✨ feat(quantity): enable non-Array quax.ArrayValue as Quantity's value (#358)
