### 2025-01-30 e90a3b7
💥 boom(quantity): rename UncheckedQuantity to BareQuantity
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
