+++ src/unxt/_src/quantity/__init__.py
@@ -4,6 +4,7 @@ __all__ = [
     "AbstractParametricQuantity",
     "AbstractQuantity",
     "Quantity",
+    "BareQuantity",
     "UncheckedQuantity",
     "is_unit_convertible",
     "uconvert",
@@ -16,5 +17,5 @@ from .api import is_unit_convertible, uconvert, ustrip
 from .base import AbstractQuantity, is_any_quantity
 from .base_parametric import AbstractParametricQuantity
 from .quantity import Quantity
-from .unchecked import UncheckedQuantity
+from .unchecked import BareQuantity, UncheckedQuantity
 from .value import convert_to_quantity_value
