+++ src/unxt/_src/quantity/register_primitives.py
@@ -60,12 +60,12 @@ def abs_p(x: AbstractQuantity) -> AbstractQuantity:
     >>> abs(q)
     Quantity['length'](Array(1, dtype=int32, ...), unit='m')
 
-    >>> from unxt.quantity import UncheckedQuantity
-    >>> q = UncheckedQuantity(-1, "m")
+    >>> from unxt.quantity import BareQuantity
+    >>> q = BareQuantity(-1, "m")
     >>> jnp.abs(q)
-    UncheckedQuantity(Array(1, dtype=int32, ...), unit='m')
+    BareQuantity(Array(1, dtype=int32, ...), unit='m')
     >>> abs(q)
-    UncheckedQuantity(Array(1, dtype=int32, ...), unit='m')
+    BareQuantity(Array(1, dtype=int32, ...), unit='m')
 
     """
     return replace(x, value=qlax.abs(ustrip(x)))
@@ -81,10 +81,10 @@ def acos_p_aq(x: AbstractQuantity) -> AbstractQuantity:
     Examples
     --------
     >>> import quaxed.numpy as xp
-    >>> from unxt.quantity import UncheckedQuantity
-    >>> q = UncheckedQuantity(-1, "")
+    >>> from unxt.quantity import BareQuantity
+    >>> q = BareQuantity(-1, "")
     >>> jnp.acos(q)
-    UncheckedQuantity(Array(3.1415927, dtype=float32, ...), unit='rad')
+    BareQuantity(Array(3.1415927, dtype=float32, ...), unit='rad')
 
     >>> from unxt.quantity import Quantity
     >>> q = Quantity(-1, "")
@@ -106,10 +106,10 @@ def acosh_p_aq(x: AbstractQuantity) -> AbstractQuantity:
     Examples
     --------
     >>> import quaxed.numpy as xp
-    >>> from unxt.quantity import UncheckedQuantity
-    >>> q = UncheckedQuantity(2.0, "")
+    >>> from unxt.quantity import BareQuantity
+    >>> q = BareQuantity(2.0, "")
     >>> jnp.acosh(q)
-    UncheckedQuantity(Array(1.316958, dtype=float32, ...), unit='rad')
+    BareQuantity(Array(1.316958, dtype=float32, ...), unit='rad')
 
     >>> from unxt.quantity import Quantity
     >>> q = Quantity(2.0, "")
@@ -132,14 +132,14 @@ def add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     Examples
     --------
     >>> import quaxed.numpy as jnp
-    >>> from unxt.quantity import UncheckedQuantity
+    >>> from unxt.quantity import BareQuantity
 
-    >>> q1 = UncheckedQuantity(1, "km")
-    >>> q2 = UncheckedQuantity(500.0, "m")
+    >>> q1 = BareQuantity(1, "km")
+    >>> q2 = BareQuantity(500.0, "m")
     >>> jnp.add(q1, q2)
-    UncheckedQuantity(Array(1.5, dtype=float32, ...), unit='km')
+    BareQuantity(Array(1.5, dtype=float32, ...), unit='km')
     >>> q1 + q2
-    UncheckedQuantity(Array(1.5, dtype=float32, ...), unit='km')
+    BareQuantity(Array(1.5, dtype=float32, ...), unit='km')
 
     >>> from unxt.quantity import Quantity
     >>> q1 = Quantity(1, "km")
@@ -149,7 +149,7 @@ def add_p_aqaq(x: AbstractQuantity, y: AbstractQuantity) -> AbstractQuantity:
     >>> q1 + q2
[…]
