### 2025-01-30 bd07b85
docs(readme): add codspeed badge (#387)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
