### 2025-01-30 841a128
✨ feat(quantity): add promotion rule (#382)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
