### 2025-02-04 998e682
🐛 fix(quantity): concat_p for abstract quantity (#389)
