+++ src/unxt/_src/quantity/register_primitives.py
@@ -1020,10 +1020,8 @@ def concatenate_p_qnd(
 
 @register(lax.concatenate_p)
 def concatenate_p_vqnd(
-    operand0: ArrayLike,
-    *operands: AbstractParametricQuantity["dimensionless"],
-    dimension: Any,
-) -> AbstractParametricQuantity["dimensionless"]:
+    operand0: ArrayLike, *operands: AbstractQuantity, dimension: Any
+) -> AbstractQuantity:
     """Concatenate quantities and arrays with dimensionless units.
 
     Examples
@@ -1045,16 +1043,8 @@ def concatenate_p_vqnd(
                   unit='')
 
     """
-    return Quantity(
-        lax.concatenate(
-            [
-                (ustrip(one, op) if hasattr(op, "unit") else op)
-                for op in (operand0, *operands)
-            ],
-            dimension=dimension,
-        ),
-        unit=one,
-    )
+    arrs = [operand0, *(ustrip(one, op) for op in operands)]
+    return Quantity(lax.concatenate(arrs, dimension=dimension), unit=one)
 
 
 # ==============================================================================
