### 2025-02-04 9175db2
🐛 fix(quantity): concat_p for abstract quantity (#389)
(cherry picked from commit 998e682c2b27243a5ebdd12f11a98dd4275e3756)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
