### 2025-02-05 bf62bbe
💥 boom(quantity): rename UncheckedQuantity to BareQuantity (#386)
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
