### 2025-02-05 447f25e
Backport PR #391: 🐛 fix(quantity): hash method for immutability checks
