### 2025-02-05 d02fffe
Backport PR #391: 🐛 fix(quantity): hash method for immutability checks (#392)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
