### 2025-02-07 2a947ca
➕ dep-add(quax-blocks): replace quaxed.experimental with quax-blocks v0.1
