### 2025-02-07 fcb20bd
🔧 config(sybil): update test configuration (#394)
