+++ pyproject.toml
@@ -36,7 +36,7 @@
     "optional-dependencies>=0.3.2",
     "plum-dispatch>=2.5.6",
     "quax>=0.0.5",
-    "quax-blocks>=0.1",
+    "quax-blocks>=0.2",
     "quaxed>=0.8.1",
     "xmmutablemap>=0.1",
     "zeroth>=1.0.0",
