### 2025-02-07 572290e
⬆️ dep-bump(quax-blocks): v0.2+ (#396)
