### 2025-02-07 8b61f34
✨ feat(quantity): AllowValue flag to let ustrip work on non-quantities
This assumes the value to be in the right unit.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
