+++ src/unxt/quantity.py
@@ -27,6 +27,7 @@ __all__ = [
     "is_unit_convertible",
     "is_any_quantity",
     "convert_to_quantity_value",
+    "AllowValue",
 ]
 
 
@@ -38,18 +39,19 @@ with install_import_hook("unxt.quantity", RUNTIME_TYPECHECKER):
     from ._src.quantity.api import is_unit_convertible, uconvert, ustrip
     from ._src.quantity.base import AbstractQuantity, is_any_quantity
     from ._src.quantity.base_parametric import AbstractParametricQuantity
+    from ._src.quantity.flag import AllowValue
     from ._src.quantity.quantity import Quantity
     from ._src.quantity.unchecked import BareQuantity, UncheckedQuantity
     from ._src.quantity.value import convert_to_quantity_value
 
-# isort: split
-# Register dispatches and conversions
-from ._src.quantity import (
-    register_api,
-    register_conversions,
-    register_dispatches,
-    register_primitives,
-)
+    # isort: split
+    # Register dispatches and conversions
+    from ._src.quantity import (
+        register_api,
+        register_conversions,
+        register_dispatches,
+        register_primitives,
+    )
 
 # Clean up namespace
 del register_conversions, register_api, register_dispatches, register_primitives
