### 2025-02-11 #7583 @ngoldbaum
Maybe @tbekolay would be interested in reviewing this paper? I enjoyed their talk about unit libraries at SciPy 2013: https://youtu.be/N-edLdxiM40?si=wqz1bIYaQv-YyMQx
