### 2025-02-11 #7583 @tbekolay
Hi, yes I think I can spare a few cycles to do this review. It will be my first for JOSS so it may take me a bit to figure out the process, please feel free to correct me if I've put anything in the wrong place :+1:
