+++ .github/workflows/ci.yml
@@ -151,7 +151,7 @@ jobs:
 
   benchmark:
     name: Run Benchmarks
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-24.04
     needs: [format]
     if:
       github.event_name == 'workflow_dispatch' || (github.event_name ==
