### 2025-02-17 #7771 @danielskatz
👋 @Cadair - when you get a chance, please generate your checklist. You should use the command `@editorialbot generate my checklist` to do this. @editorialbot commands need to be the first thing in a new comment.
