### 2025-02-19 bfdac93
✨ feat(quantity): allowvalue number
