### 2025-02-19 161f4d7
✨ feat(quantity): allowvalue number (#399)
