### 2025-02-23 4b5aecc
🐛 fix(quantity): AllowValue on Astropy
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
