### 2025-02-23 8693972
Update quantity.py
