### 2025-02-23 7396e6d
🐛 fix(quantity): AllowValue on Astropy (#400)
* 🐛 fix(quantity): AllowValue on Astropy

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
