### 2025-02-26 49b154b
🩹 fix-simple(dimensions): add missing dispatches for dimension_of
