+++ pyproject.toml
@@ -27,7 +27,7 @@
   license.file = "LICENSE"
   dependencies = [
     "astropy>=7.0.0",
-    "dataclassish>=0.5",
+    "dataclassish>=0.8.0",
     "equinox>=0.11.8",
     "is-annotated>=1.0",
     "jax>0.4.32",
