### 2025-02-26 8709c16
🩹 fix-simple(dimensions): add missing dispatches for dimension_of (#401)
