### 2025-03-11 #7771 @Cadair
Hi @nstarman, this package looks super useful (more things with units is always better), I haven't used JAX much so I can't comment on the API design or anything but it's well built, maintained and documented. Just a couple of small points:

> **Contribution and authorship:** Has the submitting author ([@nstarman](https://github.com/nstarman)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

It looks like there are a couple of people in the git history who aren't authors on the paper and one on the paper who isn't in the git history? I just wanted to confirm that people weren’t missing?

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

This seems to be missing. There's a cursory "we welcome contributions" at the end of the readme, but maybe a `CONTRIBUTING.md` and/or a section in the documentation would be good, especially encouraging people to open issues.

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

fwiw, I tried to run nox locally and it complained it couldn't find pytest when trying to run the tests, despite also claiming to have installed it. It's probably some weird interaction with my venvs / uv / nox, but thought it was worth mentioning.
