### 2025-03-15 #7771 @danielskatz
👋 @Cadair & @tbekolay - it looks like this review is moving along.

I see some specific questions/needs from @Cadair [above](https://github.com/openjournals/joss-reviews/issues/7771#issuecomment-2713437356).

@tbekolay - I see you have checked off a bunch of items, but not others. What do you need in order to check off the others?
