### 2025-03-16 3f84673
📝 docs(joss): add section on contributing
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
