### 2025-03-16 #7771 @nstarman
> Hi @nstarman, this package looks super useful (more things with units is always better), I haven't used JAX much so I can't comment on the API design or anything but it's well built, maintained and documented.

Thanks!

> It looks like there are a couple of people in the git history who aren't authors on the paper and one on the paper who isn't in the git history? I just wanted to confirm that people weren’t missing?

Yes, the paper authorship is correct. We (the authors) thought about, met weekly, and discussed the API, even if not everyone contributed specific lines of code.

> This seems to be missing. There's a cursory "we welcome contributions" at the end of the readme, but maybe a CONTRIBUTING.md and/or a section in the documentation would be good, especially encouraging people to open issues.

That sounds good. On it.

> fwiw, I tried to run nox locally and it complained it couldn't find pytest when trying to run the tests, despite also claiming to have installed it. It's probably some weird interaction with my venvs / uv / nox, but thought it was worth mentioning.

Maybe try ``uv run --group test nox -s tests``?
