### 2025-03-26 #7771 @Cadair
Yes, I am happy. Me not quite wrapping my head around the tests locally isn't an issue.
