### 2025-03-26 ea04347
minor changes in paper.md
Signed-off-by: Daniel S. Katz <d.katz@ieee.org>
