### 2025-03-26 d51b083
minor changes in paper.bib
Signed-off-by: Daniel S. Katz <d.katz@ieee.org>
