+++ paper/paper.bib
@@ -18,7 +18,7 @@
   month         = oct,
   pages         = {A33},
   primaryclass  = {astro-ph.IM},
-  title         = {{Astropy: A community Python package for astronomy}},
+  title         = {Astropy: A community {P}ython package for astronomy},
   volume        = {558},
   year          = 2013
 }
@@ -37,7 +37,7 @@
   number        = {2},
   pages         = {167},
   primaryclass  = {astro-ph.IM},
-  title         = {{The Astropy Project: Sustaining and Growing a Community-oriented Open-source Project and the Latest Major Release (v5.0) of the Core Package}},
+  title         = {The {A}stropy Project: Sustaining and Growing a Community-oriented Open-source Project and the Latest Major Release (v5.0) of the Core Package},
   volume        = {935},
   year          = 2022
 }
@@ -60,7 +60,6 @@
 
 @misc{nasa:98,
   author       = {NASA},
-  howpublished = {\url{https://science.nasa.gov/mission/mars-climate-orbiter/}},
   title        = {{M}ars {C}limate {O}rbiter - {N}{A}{S}{A} {S}cience --- science.nasa.gov},
   url          = {https://science.nasa.gov/mission/mars-climate-orbiter/},
   year         = {}
@@ -92,7 +91,8 @@
 @software{pint,
   author = {Grecco, Hernan},
   date   = {2012},
-  title  = {Pint: makes units easy}
+  title  = {Pint: makes units easy},
+  url = {https://github.com/hgrecco/pint},
 }
 
 @software{quax:23,
@@ -110,7 +110,7 @@
   number    = {28},
   pages     = {809},
   publisher = {The Open Journal},
-  title     = {unyt: Handle,  manipulate,  and convert data with units in Python},
+  title     = {{u}nyt: Handle,  manipulate,  and convert data with units in {P}ython},
   url       = {https://doi.org/10.21105/joss.00809},
   volume    = {3},
   year      = {2018}
