### 2025-03-26 #7771 @danielskatz
👋 @nstarman - I've suggested some small changes in https://github.com/GalacticDynamics/unxt/pull/403 - please merge this or let me know what you disagree with.

After that, please:
