### 2025-03-28 c26e696
ci: remove JOSS paper builder
paper is finalized.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
