+++ .github/workflows/paper-pdf.yml
@@ -1,32 +1,0 @@
-name: JOSS Paper PDF
-on:
-  push:
-    paths:
-      - paper/**
-      - .github/workflows/draft-pdf.yml
-
-permissions:
-  contents: write
-
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: JOSS Paper
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          paper-path: paper/paper.md
-      - name: Upload PDF artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: paper
-          path: paper/paper.pdf
-      - name: Commit PDF to repository
-        uses: EndBug/add-and-commit@v9
-        with:
-          message: "(auto) Paper PDF"
-          add: "paper/*.pdf"
