### 2025-03-28 0386a85
docs: JOSS paper (#404)
* 📝 docs(joss): paper
* 📝 docs(joss): add section on related works
* 📝 docs(joss): add section on contributing
* minor changes for JOSS publication (#403)
* minor changes in paper.md
* minor changes in paper.bib
* ci: remove JOSS paper builder

Co-authored-by: Daniel S. Katz <d.katz@ieee.org>
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
