+++ paper/paper.bib
@@ -1,0 +1,117 @@
+@misc{array-api,
+  author = {Array API Consortion},
+  title  = {Array API Standard},
+  url    = {https://data-apis.org/array-api/latest/index.html},
+  year   = {2021}
+}
+
+@article{astropy:13,
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..33A},
+  archiveprefix = {arXiv},
+  author        = {{Astropy Collaboration} and {Robitaille}, Thomas P. and {Tollerud}, Erik J. and {Greenfield}, Perry and {Droettboom}, Michael and {Bray}, Erik and {Aldcroft}, Tom and {Davis}, Matt and {Ginsburg}, Adam and {Price-Whelan}, Adrian M. and {Kerzendorf}, Wolfgang E. and {Conley}, Alexander and {Crighton}, Neil and {Barbary}, Kyle and {Muna}, Demitri and {Ferguson}, Henry and {Grollier}, Fr{\'e}d{\'e}ric and {Parikh}, Madhura M. and {Nair}, Prasanth H. and {Unther}, Hans M. and {Deil}, Christoph and {Woillez}, Julien and {Conseil}, Simon and {Kramer}, Roban and {Turner}, James E.~H. and {Singer}, Leo and {Fox}, Ryan and {Weaver}, Benjamin A. and {Zabalza}, Victor and {Edwards}, Zachary I. and {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, Andrew R. and {Crawford}, Steven M. and {Dencheva}, Nadia and {Ely}, Justin and {Jenness}, Tim and {Labrie}, Kathleen and {Lim}, Pey Lian and {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
+  doi           = {10.1051/0004-6361/201322068},
+  eid           = {A33},
+  eprint        = {1307.6212},
+  journal       = {\aap},
+  keywords      = {methods: data analysis, methods: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
+  month         = oct,
+  pages         = {A33},
+  primaryclass  = {astro-ph.IM},
+  title         = {Astropy: A community {P}ython package for astronomy},
+  volume        = {558},
+  year          = 2013
+}
+
+@article{astropy:22,
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2022ApJ...935..167A},
+  archiveprefix = {arXiv},
[…]
