+++ docs/index.md
@@ -39,6 +39,7 @@ api/index.md
 
 glossary.md
 conventions.md
+contributing.md
 ```
 
 # 🚀 Get Started
