### 2025-03-28 #7771 @nstarman
Thanks @danielskatz and reviewers @Cadair @tbekolay!

- I made a tagged release — https://github.com/GalacticDynamics/unxt/releases/tag/v1.4.0
- Archived in Zenodo with DOI [10.5281/zenodo.15103023](https://doi.org/10.5281/zenodo.15103023) and checked the title and author list are adjusted.
