### 2025-03-28 #7771 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
